maserver/requirements.txt
python-dotenv
PyYAML
structlog
filelock
binance
psutil